requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncpg>=0.29.0
pydantic-settings>=2.5.0
loguru>=0.7.0

SQLAlchemy>=2.0.43
alembic>=1.16.0
azure-core>=1.35.0
pydantic>=2.11.7
pydantic_core>=2.33.2

opentelemetry-api >= 1.10.0
opentelemetry-sdk >= 1.10.0
azure-monitor-opentelemetry >= 1.1.1